Fix RX scanning loop
One of the functions returned void from map which RX evaluated as null and broke the chain.

Solution:
We should either always make sure to return something other than null from the map.
Or in this case I opted for conversion of the maps into side effects (doOnNext) as it's not
actaually part of the chain/flow