Make sure UNKNOWN iOS devices are asked for BUID again (#85)
* Make sure UNKNOWN iOS devices are asked for BUID again

I noticed that sometimes an iOS devices that failed to send it's BUID over GATT would
stay as UNKNOWN forever. It looks like that calling .close() on the GATT connection
does then in turn skip the onDisconnected callback and that is where we were clearing
the ScanSession.
I changed it so that it actually always sets the timer when disconnecting plus I added a more
aggresssive clearing when scanning is restarted, so that all UNKNOWN iOS devices are
purged from the cache.