Migrates from BUID broadcasting to TUID broadcasting (#103)
Background:
Put simply we wanted to make it more secure and private. Periodic rotation of the broadcasted identifier
makes it virtually impossible to track the user over time.

Changes:
The client now receives also a list of TUIDs upon registration and saves those as a set to shared prefs.
Then when broadcasting starts it randomly picks one of those and starts broadcasting it. When broadcasting
gets restarted after a set interval new TUID is chosen from the set. We also never choose the same TUID twice in a row.

- Sandbox now also shows the currently broadcasted TUID
- My data screen doesn't show the bumber of encounters as we are now actually no longer able to track this locally.