Don't trigger notification when screen goes on&off with battery saver on
Problem:
Toggling screen state also toggles location state with battery saver on. We listen for
location services state changes so that was what was triggering the notifications.

Change:
Check if battery saver is enabled when the location state broadcast is received.
We can swallow the event in that case as the app should already be in a disabled/error state.