Remove wakelock
This removes the need for wakelock which should improve battery consumption.
Because of that the way how scans are saved has to be changed.
It now uses timestamp from scan results directly because thosee results can come delayed and batched.
It also then splits the results into buckets where each is 2 minutes long.